+++ code/crates/sync/src/handle.rs
@@ -58,11 +58,12 @@ pub enum Input<Ctx: Context> {
     /// A status update has been received from a peer
     Status(Status<Ctx>),
 
-    /// Consensus just started a new height
-    StartHeight(Ctx::Height),
+    /// Consensus just started a new height.
+    /// The boolean indicates whether this was a restart or a new start.
+    StartedHeight(Ctx::Height, bool),
 
     /// Consensus just decided on a new value
-    UpdateHeight(Ctx::Height),
+    Decided(Ctx::Height),
 
     /// A ValueSync request has been received from a peer
     ValueRequest(InboundRequestId, PeerId, ValueRequest<Ctx>),
@@ -106,28 +107,36 @@ where
 
         Input::Status(status) => on_status(co, state, metrics, status).await,
 
-        Input::StartHeight(height) => on_start_height(co, state, metrics, height).await,
+        Input::StartedHeight(height, restart) => {
+            on_started_height(co, state, metrics, height, restart).await
+        }
 
-        Input::UpdateHeight(height) => on_update_height(co, state, metrics, height).await,
+        Input::Decided(height) => on_decided(co, state, metrics, height).await,
 
         Input::ValueRequest(request_id, peer_id, request) => {
             on_value_request(co, state, metrics, request_id, peer_id, request).await
         }
+
         Input::ValueResponse(request_id, peer_id, response) => {
             on_value_response(co, state, metrics, request_id, peer_id, response).await
         }
+
         Input::GotDecidedValue(request_id, height, value) => {
             on_value(co, state, metrics, request_id, height, value).await
         }
+
         Input::SyncRequestTimedOut(peer_id, request) => {
             on_sync_request_timed_out(co, state, metrics, peer_id, request).await
         }
+
         Input::InvalidCertificate(peer, certificate, error) => {
             on_invalid_certificate(co, state, metrics, peer, certificate, error).await
         }
+
         Input::GetVoteSet(height, round) => {
             on_get_vote_set(co, state, metrics, height, round).await
         }
+
         Input::VoteSetRequest(request_id, peer_id, request) => {
             on_vote_set_request(co, state, metrics, request_id, peer_id, request).await
         }
@@ -142,7 +151,6 @@ where
     }
 }
 
-#[tracing::instrument(skip_all)]
 pub async fn on_tick<Ctx>(
     co: Co<Ctx>,
     state: &mut State<Ctx>,
@@ -158,13 +166,6 @@ where
     Ok(())
 }
 
-#[tracing::instrument(
-    skip_all,
-    fields(
-        height.sync = %state.sync_height,
-        height.tip = %state.tip_height
-    )
-)]
 pub async fn on_status<Ctx>(
     co: Co<Ctx>,
     state: &mut State<Ctx>,
@@ -174,9 +175,9 @@ pub async fn on_status<Ctx>(
 where
     Ctx: Context,
 {
-    debug!(%status.peer_id, %status.height, "Received peer status");
+    debug!(%status.peer_id, %status.tip_height, "Received peer status");
 
-    let peer_height = status.height;
+    let peer_height = status.tip_height;
 
     state.update_status(status);
 
@@ -201,7 +202,48 @@ where
     Ok(())
 }
 
-#[tracing::instrument(skip_all)]
+pub async fn on_started_height<Ctx>(
+    co: Co<Ctx>,
+    state: &mut State<Ctx>,
+    metrics: &Metrics,
+    height: Ctx::Height,
+    restart: bool,
+) -> Result<(), Error<Ctx>>
+where
+    Ctx: Context,
+{
+    let tip_height = height.decrement().unwrap_or(height);
+
+    debug!(height.tip = %tip_height, height.sync = %height, %restart, "Starting new height");
+
+    state.started = true;
+    state.sync_height = height;
+    state.tip_height = tip_height;
+
+    // Check if there is any peer already at or above the height we just started,
+    // and request sync from that peer in order to catch up.
+    request_value(co, state, metrics).await?;
+
+    Ok(())
+}
+
+pub async fn on_decided<Ctx>(
+    _co: Co<Ctx>,
+    state: &mut State<Ctx>,
+    _metrics: &Metrics,
+    height: Ctx::Height,
+) -> Result<(), Error<Ctx>>
+where
+    Ctx: Context,
+{
+    debug!(height.tip = %height, "Updating tip height");
+
+    state.tip_height = height;
+    state.remove_pending_decided_value_request(height);
+
+    Ok(())
+}
+
 pub async fn on_value_request<Ctx>(
     co: Co<Ctx>,
     _state: &mut State<Ctx>,
@@ -222,7 +264,6 @@ where
     Ok(())
 }
 
-#[tracing::instrument(skip_all)]
 pub async fn on_value_response<Ctx>(
     _co: Co<Ctx>,
     state: &mut State<Ctx>,
@@ -239,46 +280,6 @@ where
     state.remove_pending_decided_value_request(response.height);
 
     metrics.decided_value_response_received(response.height.as_u64());
-
-    Ok(())
-}
-
-pub async fn on_start_height<Ctx>(
-    co: Co<Ctx>,
-    state: &mut State<Ctx>,
-    metrics: &Metrics,
-    height: Ctx::Height,
-) -> Result<(), Error<Ctx>>
-where
-    Ctx: Context,
-{
-    debug!(height.sync = %height, "Starting new height");
-
-    state.started = true;
-    state.sync_height = height;
-
-    // Check if there is any peer already at or above the height we just started,
-    // and request sync from that peer in order to catch up.
-    request_value(co, state, metrics).await?;
-
-    Ok(())
-}
-
-pub async fn on_update_height<Ctx>(
-    _co: Co<Ctx>,
-    state: &mut State<Ctx>,
-    _metrics: &Metrics,
-    height: Ctx::Height,
-) -> Result<(), Error<Ctx>>
-where
-    Ctx: Context,
-{
-    if state.tip_height < height {
-        debug!(height.tip = %height, "Updating tip height");
-
-        state.tip_height = height;
-        state.remove_pending_decided_value_request(height);
-    }
 
     Ok(())
 }
@@ -369,10 +370,10 @@ where
         return Ok(());
     }
 
-    if let Some(peer) = state.random_peer_with_value(sync_height) {
+    if let Some(peer) = state.random_peer_with_tip_at_or_above(sync_height) {
         request_value_from_peer(co, state, metrics, sync_height, peer).await?;
     } else {
-        warn!(height.sync = %sync_height, "No peer to request sync from");
+        debug!(height.sync = %sync_height, "No peer to request sync from");
     }
 
     Ok(())
@@ -418,7 +419,7 @@ where
     info!(height.sync = %certificate.height, "Requesting sync from another peer");
     state.remove_pending_decided_value_request(certificate.height);
 
-    let Some(peer) = state.random_peer_with_value_except(certificate.height, from) else {
+    let Some(peer) = state.random_peer_with_tip_at_or_above_except(certificate.height, from) else {
         error!(height.sync = %certificate.height, "No other peer to request sync from");
         return Ok(());
     };
@@ -441,7 +442,7 @@ where
         return Ok(());
     }
 
-    let Some(peer) = state.random_peer_for_votes(height, round) else {
+    let Some(peer) = state.random_peer_with_sync_at(height, round) else {
         warn!(%height, %round, "No peer to request vote set from");
         return Ok(());
     };
@@ -476,7 +477,6 @@ where
     Ok(())
 }
 
-#[tracing::instrument(skip_all)]
 pub async fn on_vote_set_request<Ctx>(
     _co: Co<Ctx>,
     _state: &mut State<Ctx>,
@@ -516,7 +516,6 @@ where
     Ok(())
 }
 
-#[tracing::instrument(skip_all)]
 pub async fn on_vote_set_response<Ctx>(
     _co: Co<Ctx>,
     state: &mut State<Ctx>,