+++ code/crates/test/tests/it/main.rs
@@ -3,6 +3,7 @@ mod n3f0;
 mod n3f0_consensus_mode;
 mod n3f0_pubsub_protocol;
 mod n3f1;
+mod reset;
 mod value_sync;
 mod vote_sync;
 mod vote_sync_bcast;